Added some settings
It is now possible to select the sites which should be searched.
I also added Checkboxes to send Debug Information to me and to autofill unfilled Textfields.

None of the Settings have any effect at the moment, but they get saved for later usage.

Also improved the search to be more efficient and made it more dynamic.